requirements.txt
# --- Core framework ---
fastapi==0.104.1
uvicorn[standard]==0.24.0

# --- Database & ORM ---
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # Driver PostgreSQL para la nube

# --- Security & auth ---
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4

# --- Utilities ---
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
requests==2.31.0

